corrected deprecated isless

+++ src/parser.jl
@@ -59,7 +59,7 @@ nameStart(c::Char) = 'a' <= c && c <= 'z' || 'A' <= c && c <= 'Z' || c == '_' ||
 
 # // nameChar returns whether c can be a character within an identifier
 # // (not counting an escape sequence).
-nameChar(c::Char) = 'a' <= c && c <= 'z' || 'A' <= c && c <= 'Z' || c == '_' || c > 127 ||
+nameChar(c::Char) = 'a' <= c && c <= 'z' || 'A' <= c && c <= 'Z' || c == '_' || UInt32(c) > 127 ||
 	c == '-' || '0' <= c && c <= '9'
